chore: rename getExpectedFieldData to getFieldData
